Use DSFR as a dependency instead of vendored
Benefits:
- Less files in our repo
- Dependabot can tell us when a new version is available
- Less HTTP requests for the CSS (1 instead of 3)
- The files can be optimised by Parcel (751KB instead of 819KB of CSS)

We're only keeping the favicon as it is difficult to use through NPM and Parcel.

We use `"errorRecovery": true` to workaround some [bad CSS in DSFR][]

By compiling the DSFR, our local development server takes longer to start if we delete the previous Parcel builds. That's why we're removing the automatic deletions in `package.json`.

fix https://github.com/betagouv/SPPNautInterface/issues/7

[bad CSS in DSFR]: https://github.com/GouvernementFR/dsfr/issues/576