One boarding place has at most one district (#311)
We enforce it through:
- The PilotageDistrict <-> PilotService connection via `PilotService.clean()`
- The PilotService <-> PilotBoardingPlace connection via `PilotBoardingPlaceServiceThrough.clean()`. `PilotBoardingPlaceServiceThrough` is the same as the implicit ManyToMany model Django created, except for the table and field names.

`PilotBoardingPlace` gets a `pilotage_district` property thanks to this enforcement.

The `state_operations` of the migration are the same as the automated one created by Django. We wrote the `database_operations` to keep the data around. Inspired by https://docs.djangoproject.com/en/4.2/howto/writing-migrations/#changing-a-manytomanyfield-to-use-a-through-model

---------

Co-authored-by: Anthony Ricaud <anthony@ricaud.me>